renamed helpers to utilities to match naming convention in repo, differentiated setup and smoke projects in playwright config, renamed test spec file to a more descriptive name
